Add trusted dealer to key-share itself
