+++ key-share/Cargo.toml
@@ -13,6 +13,7 @@ keywords = ["mpc", "threshold-signatures", "tss"]
 [dependencies]
 generic-ec = "0.1.3"
 generic-ec-zkp = "0.1"
+rand_core = { version = "0.6", optional = true }
 
 slip-10 = { version = "0.1", optional = true, features = ["std"] }
 udigest = { version = "0.1", default-features = false, features = ["derive"], optional = true }
@@ -23,7 +24,11 @@ serde_with = { version = "2", optional = true}
 
 thiserror = "1"
 
+[dev-dependencies]
+rand_core = { version = "0.6", features = ["getrandom"] }
+
 [features]
 serde = ["dep:serde", "serde_with", "hex"]
 hd-wallets = ["slip-10"]
+spof = ["dep:rand_core"]
 udigest = ["dep:udigest"]
